+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added support for extension-provided skills
 - using baseCommit when rebasing the worktree on different branch
 - added base branch name to the worktree merge button and improved rebasing
+- hide save prompt button when editing last message
 
 ## [0.63.0]